Added Windows support. (But there's a BUG: 'enter' not working on client side) Some Fixes in CMakeLists No uint datatype in MSVC, therefore: uint -> uint64 Explicit include <iterator>, instead of implicit.

+++ include/server_headers/Chatroom.hpp
@@ -33,7 +33,7 @@ namespace meow {
         private:
 //            const tcp::socket to_server_socket_;
             tcp::socket chatroom_socket_;
-            static const uint max_recent_msgs;
+            static const uint64_t max_recent_msgs;
 
             std::set <chat_participant_ptr> participants_;
             chat_message_queue recent_msgs_;
